+++ upgrade/patches/44_usr_relative_dates.php
@@ -21,7 +21,7 @@ $db->query('ALTER TABLE {{%user_preference}} ADD COLUMN upr_relative_date tinyin
 
 $setup = Setup::get();
 
-$db->query('UPDATE {{%user_preference}} SET upr_relative_date=?', array($setup['relative_date'] == 'enabled'));
+$db->query('UPDATE {{%user_preference}} SET upr_relative_date=?', array((int) ($setup['relative_date'] == 'enabled')));
